refactor: 주문 생성 및 정보 저장 로직 리팩토링 - Parser 구현으로 사용하지 않는 클래스 제거
