+++ src/main/java/christmas/Application.java
@@ -2,11 +2,10 @@ package christmas;
 
 import christmas.controller.Controller;
 import christmas.service.EventService;
-import christmas.service.OrderService;
 
 public class Application {
     public static void main(String[] args) {
-        Controller controller = new Controller(new OrderService(), new EventService());
+        Controller controller = new Controller(new EventService());
         controller.run();
     }
 }
