--- src/main/java/christmas/utils/validator/ValidationUtils.java
@@ -1,11 +1,0 @@
-package christmas.utils.validator;
-
-public class ValidationUtils {
-    public static int parseInteger(String input, String errorMessage) {
-        try {
-            return Integer.parseInt(input.trim());
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException(errorMessage);
-        }
-    }
-}
